register: simplify/fix typing of float registers
The encode function is broken for float registers as the value is
converted to int before being passed to struct.pack (as required
for the integer types). Fix this by making the proper type a class
attribute. This also allows the init function of Reg_f32l to be
dropped.