carlo_gavazzi: Always format versions as text
There is currently rules in gui-v2 for formatting FirmwareVersion,
for services that don't follow the normal rules. The same issue
however shows up for the HardwareVersion. So instead of adding
more exceptions to gui-v2, rather just format the version as a text
string. This makes it unambigious for all UIs, including VRM and
GUI-v2.

https://github.com/victronenergy/venus/issues/1464